requirements.txt
streamlit
python_dotenv
pycountry
huggingface_hub
transformers
torch
torchvision
requests